recipe: Add Browser install alternatives
Epiphany Browser from Ubuntu is much buggy and videos not working without
ffmpeg preinstalled in base image
All the alternatives is open source
Chromium: Snap and Flatpak.
Firefox: Snap and Flatpak.
Brave: Snap and Flatpak.

Bugs: All Flatpak packages for someone reason the VAAPI is not working
in the minimal cases just not working in Wayland, and the maintainers not
looking in try to fix these problems.
Just the Snap Packages have a Good VAAPI compatibility support.