fix:[close #209] User does not always have permissions for changes
tested both on bare metal and vm, first setup does not prompt for the password.

Side note: since this update will not be in the ISO immediately, the processor will replace it with the one in upstream. Be sure to replace it again on a tty before logging in again, otherwise it will ask for a password.