Changed the testing system
Now it will use a JSON file called tests.json. The tests (and which options have to be selected for those tests) will be stored in that file. The old system did not work for a few reasons:
1. It used commands to verify that the package was installed but most package did not have binary files installed. Instead it now checks for the file being present.
2. It used to check for the commands being present during the same boot that they were installed in. Instead it now checks on the next boot.
The tests run when First Setup starts and if it's the next boot the tests run to check if the files installed by the packages are present. If they are, it proceeds with the next boot. if they are not, it run the first boot again (meaning the packages are installed again)